+++ CHANGELOG.md
@@ -36,6 +36,9 @@
 * Optimize `page.matching`. Push the matching down the tree, as long
   as the join clause does not use the new image attribute.
 
+* Optimize `autowrap.matching`. Push the matching down the tree, as
+  long as the join clause only uses untouched attributes.
+
 # 0.18.0 - 2021/03/12
 
 * Default Relation#type is provided, that returns Bmg::Type::ANY